tests/wpt/tests/css/compositing/parsing/isolation-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Compositing and Blending Level 1: getComputedStyle().isolation</title>
<link rel="help" href="https://drafts.fxtf.org/compositing-1/#propdef-isolation">
<meta name="assert" content="isolation computed value is as specified.">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("isolation", "auto");
test_computed_value("isolation", "isolate");
</script>
</body>
</html>
